Adk_Copilotkit_UI_App/backend/pyproject.toml
[project]
name = "adk-copilotkit-backend"
version = "0.1.0"
description = "ADK Python backend with Deal Builder and Knowledge Q&A agents, Postgres sessions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
    "google-adk>=1.0.0",
    "ag-ui-adk>=0.3.0",
    "asyncpg>=0.28.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "httpx"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
